remove dolt sql-client command
