--- integration-tests/bats/sql-client-no-privs-no-mysql.expect
@@ -1,20 +1,0 @@
-#!/usr/bin/expect
-
-# start client
-set timeout 1
-spawn dolt sql-client --host=0.0.0.0 --port=$PORT -d
-
-# list all users
-expect {
-    "mysql> " { send "select user from mysql.user;\r"; }
-}
-
-# look for only root user
-expect {
-    "root"
-}
-
-# quit
-expect {
-    "mysql> " { exit 0 }
-}
